Overwrite Klippfiles with -f. Opens the result in TextMate or TextEdit.

+++ lib/project.rb
@@ -18,8 +18,25 @@ module Project
     template = params.shift_argument
     raise 'Template name required for project init' unless template
 
-    spec = Template::Spec.from_file(Template.path_for_template template)
+    spec = Template::Spec.from_file Template.path_for_template(template)
+    filename = 'Klippfile'
+
+    force = params.splice_option('-f')
+    will_overwrite = File.exists?(filename) && force
+
+    raise "#{filename} already exists, not overwriting. Use -f to force overwriting" if File.exists?(filename) && !force
+
     File.write('Klippfile', spec.klippfile)
+
+    Formatador.display_line("[green][√] #{will_overwrite ? "Re-saved" : "Saved"} #{filename}[/]")
+
+    capture_stdout {
+      `open -a TextMate #{filename} 2>&1` if File.exists?(filename)
+    }
+
+    if $? && $?.exitstatus > 0
+      `open -t #{filename}` if File.exists?(filename)
+    end
   end
 
 end
